pyproject.toml
[project]
name = "YouTubeSummarizer"
version = "0.1.0"
description = "A Python application to summarize YouTube videos, generate diagrams, and create quizzes using Gemini API and Gradio."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "google-generativeai>=0.7.2",
    "youtube-transcript-api>=0.6.2",
    "gradio>=4.44.0",
    "matplotlib>=3.9.2",
    "graphviz>=0.20.3",
    "python-dotenv>=1.0.1",
    "pytest>=8.3.2",
]

[tool.uv]
venv = ".venv"
python = "3.10"